Update the structure for FOL-LNN
- Thay đổi cấu trúc code cho rõ ràng.
- ol_expressions_input.json: Nơi chứa đầu vào FOL, Facts, Queries
- save_load_fol.py: Thực hiện load input từ json file
- fol_lnn.py: Chứ các hàm hỗ trợ chuyển đổi từ FOL sang format LNN
- inference.py Nhận input đầu vào sau khi được chuyển đổi sang format LNN
- main.py File thực thi chính.
