+++ fol_expressions_input.json
@@ -1,0 +1,12 @@
+{
+    "expression": [
+        "∀x,(Hinh_Vuong(x)→ Chu_Nhat(x))",
+        "∀x,(Chu_Nhat(x)→ Bon_Canh(x))",
+        "∀x,(Bon_Canh(x)∧Cap_canh_song_song(x) → Hinh_thoi(x))"
+    ],
+    "facts": {
+        "Hinh_Vuong": {"S": "True"},
+        "Cap_canh_song_song": {"S": "True"}
+    },
+    "question": ["Hinh_thoi", "Chu_Nhat"]
+}
